LayoutTests/imported/w3c/web-platform-tests/navigation-api/currententrychange-event/navigation-updateCurrentEntry-expected.txt


FAIL currententrychange fires for navigation.updateCurrentEntry() assert_equals: expected 1 but got 0
